Fix ScreenFade integer truncation losing sub-second flashes
The length calculation used integer division (duration + hold_time) / 4096,
dropping fractional seconds. A 0.5s flash produced length=0, making
blinded_time end up in the past. Changed to float division.

+++ bot_client.cpp
@@ -435,7 +435,7 @@ void BotClient_Valve_ScreenFade(void *p, int bot_index)
    static int duration;
    static int hold_time;
    //static int fade_flags;
-   int length;
+   float length;
 
    if (state == 0)
    {
@@ -456,7 +456,7 @@ void BotClient_Valve_ScreenFade(void *p, int bot_index)
    {
       state = 0;
 
-      length = (duration + hold_time) / 4096;
+      length = (duration + hold_time) / 4096.0f;
       bots[bot_index].blinded_time = gpGlobals->time + length - 2.0;
    }
    else
